+++ src/app/api/v1/companies/[companyId]/plants/route.ts
@@ -1,17 +1,14 @@
 import { NextRequest, NextResponse } from "next/server"
 import { prisma } from "@/lib/prisma"
-import { Zones } from "@prisma/client"
+import { Plants } from "@prisma/client"
 import { Params, ResponseAPI } from "@/lib/@types/types"
 
 /**
- * TODO: move this route to ??
- *
- * Handle the GET request to retrieve all zones related to a specific company
- * in the database.
+ * Handle the GET request to retrieve all plants related to a specific company
  *
  * @param {NextRequest} request - The HTTP request containing the request data.
  * @param {Params<"companyId">} params - The dynamic parameter to extract the `companyId`.
- * @returns {Promise<NextResponse>} - HTTP response with the zones related to the user.
+ * @returns {Promise<NextResponse>} - HTTP response with the plants related to the user.
  * @example
  * ```ts
  * const response = await fetch("/api/v1/companies/{companyId}/plants")
@@ -21,26 +18,21 @@ import { Params, ResponseAPI } from "@/lib/@types/types"
 export const GET = async (request: NextRequest, { params }: Params<"companyId">): Promise<NextResponse> => {
     try {
         const companyId = parseInt(params.companyId)
-        const data = await prisma.plants
-            .findMany({
-                where: {
-                    companyId,
-                },
-                include: {
-                    Zones: true,
-                },
-            })
-            .then((plants) => plants.flatMap((plant) => plant.Zones))
-        return NextResponse.json<ResponseAPI<Zones[]>>({
+        const data = await prisma.plants.findMany({
+            where: {
+                companyId,
+            },
+        })
+        return NextResponse.json<ResponseAPI<unknown>>({
             data,
             ok: true,
         })
     } catch (error) {
-        return NextResponse.json<ResponseAPI<Zones[]>>(
+        return NextResponse.json<ResponseAPI<Plants[]>>(
             {
                 data: [],
                 ok: false,
-                message: "Failed to retrieve zones",
+                message: "Failed to retrieve the data",
             },
             { status: 404 }
         )
