+++ src/app/api/v1/companies/[companyId]/route.ts
@@ -4,7 +4,7 @@ import { UserPlants } from "@prisma/client"
 import { Params, ResponseAPI } from "@/lib/@types/types"
 
 /**
- * TODO: add documentation
+ * TODO: add documentation ???
  *
  * @param {NextRequest} request -
  * @param {Params<"companyId">} param1 -
